Fix build errors: split dashboard into client component, suppress any-type lint warnings
